Update README.md
In this programming assignment, you will be building an interpreter for our SPL programming language. The grammar rules of the language and its tokens were given in Programming Assignments 1 and 2. You are required to modify the parser you have implemented for the SPL language to implement an interpreter for it. The specifications of the grammar rules are described in EBNF notations as follows.

    Prog ::= StmtList

    StmtList ::= Stmt ;{ Stmt; }

    Stmt ::= AssignStme | WriteLnStmt | IfStmt

    WriteLnStmt ::= writeln (ExprList)

    IfStmt ::= if (Expr) ‘{‘ StmtList ‘}’ [ else ‘{‘ StmtList ‘}’ ]

    AssignStmt ::= Var = Expr

    Var ::= NIDENT | SIDENT

    ExprList ::= Expr { , Expr }

    Expr ::= RelExpr [(-eq|==) RelExpr ]

    RelExpr ::= AddExpr [ ( -lt | -gt | < | > ) AddExpr ]

    AddExpr :: MultExpr { ( + | – | .) MultExpr }

    MultExpr ::= ExponExpr { ( * | / | **) ExponExpr }

    ExponExpr ::= UnaryExpr { ^ UnaryExpr }

    UnaryExpr ::= [( – | + )] PrimaryExpr

    PrimaryExpr ::= IDENT | SIDENT | NIDENT | ICONST | RCONST | SCONST | (Expr)

The following points describe the SPL programming language. These points that are related to the language syntactic rules were implemented in Programming Assigning 2. However, the points related to the SPL language semantics are required to be implemented in the SPL language interpreter. These points are: